// src/interfaces/movieInterfaces.tsx
// Generated by https://quicktype.io

export interface MoviePopular {
   page: number;
   results: ResultMoviePopular[];
   total_pages: number;
   total_results: number;
}

export interface ResultMoviePopular {
   adult: boolean;
   backdrop_path: string;
   genre_ids: number[];
   id: number;
   original_language: OriginalLanguage;
   original_title: string;
   overview: string;
   popularity: number;
   poster_path: string;
   release_date: string;
   title: string;
   video: boolean;
   vote_average: number;
   vote_count: number;
}

export enum OriginalLanguage {
   En = 'en',
   Es = 'es',
   Fr = 'fr',
}